move uploadWorkflowStatus enum to file for future imports
